requirements.txt
# requirements.txt
openai==1.12.0
python-dotenv==1.0.0
PyPDF2==3.0.1
slack-sdk==3.27.0
numpy==1.24.3
scikit-learn==1.3.0
pytest==7.4.0
black==23.7.0
isort==5.12.0
mypy==1.5.1
python-dotenv==1.0.0
typing-extensions==4.7.1
pydantic==2.3.0